Migrates from multiple Swift packages to a single NOICommunityLib Swift packages. Commons modules.
