--- EventShortClient/README.md
@@ -1,3 +1,0 @@
-# EventShortClient
-
-A description of this package.
