--- EventShortTypes/README.md
@@ -1,3 +1,0 @@
-# EventShortTypesClient
-
-A description of this package.
