+++ NOICommunity/NOICommunityLib/README.md
@@ -1,3 +1,3 @@
-# AppPreferencesClient
+# NOICommunityLib
 
 A description of this package.
