+++ NOICommunity/NOICommunityLib/Sources/EventShortClient/Interface.swift
