+++ NOICommunity/NOICommunityLib/Sources/EventShortTypesClient/Interface.swift
