+++ NOICommunity/NOICommunityLib/Sources/EventShortTypesClientLive/Live.swift
@@ -5,8 +5,10 @@
 //  Created by Matteo Matassoni on 23/02/22.
 //
 
-import Combine
 import Foundation
+import Combine
+import PascalJSONDecoder
+import DecodeEmptyRepresentable
 import EventShortTypesClient
 
 // MARK: - Private Constants
@@ -86,67 +88,4 @@ private struct EventShortTypesListRequest {
         return result.isEmpty ? nil : result
     }
 
-}
-
-private extension String {
-    func firstCharLowercased() -> String {
-        prefix(1).lowercased() + dropFirst()
-    }
-}
-
-// MARK: - JSONDecoder.KeyDecodingStrategy+PascalCase
-
-extension JSONDecoder.KeyDecodingStrategy {
-    static var convertFromPascalCase: JSONDecoder.KeyDecodingStrategy {
-        return .custom { keys -> CodingKey in
-            // keys array is never empty
-            let key = keys.last!
-            // Do not change the key for an array
-            guard key.intValue == nil else {
-                return key
-            }
-            
-            let codingKeyType = type(of: key)
-            let newStringValue = key.stringValue.firstCharLowercased()
-            
-            return codingKeyType.init(stringValue: newStringValue)!
-        }
-    }
-}
-
-// MARK: - EmptyRepresentable
-
-/// A type that has an "empty" representation.
-public protocol EmptyRepresentable {
-    static func empty() -> Self
-}
-
-// MARK: - Array+EmptyRepresentable
-
-extension Array: EmptyRepresentable {
-    public static func empty() -> Array<Element> {
-        []
-    }
-}
-
-// MARK: - Dictionary+EmptyRepresentable
-
-extension Dictionary: EmptyRepresentable {
-    public static func empty() -> Dictionary<Key, Value> {
-        [:]
-    }
-}
-
-// MARK: - KeyedDecodingContainer+EmptyRepresentable
-extension KeyedDecodingContainer {
-    public func decode<T>(
-        _ type: T.Type,
-        forKey key: KeyedDecodingContainer<K>.Key
-    ) throws -> T where T: Decodable & EmptyRepresentable {
-        if let result = try decodeIfPresent(T.self, forKey: key) {
-            return result
-        } else {
-            return T.empty()
-        }
-    }
 }
