+++ NOICommunity/NOICommunityLib/Sources/PascalJSONDecoder/String+firstCharLowercased.swift
@@ -1,0 +1,18 @@
+//
+//  String+firstCharLowercased.swift
+//  PascalJSONDecoder
+//
+//  Created by Matteo Matassoni on 11/03/22.
+//
+
+import Foundation
+
+// MARK: - String+firstCharLowercased
+
+extension String {
+
+    func firstCharLowercased() -> String {
+        prefix(1).lowercased() + dropFirst()
+    }
+
+}
